Reduce duplication by extracting ModEasyWithExtraLives
